Changed frontend to TypeScript. Added Mantine UI.
