+++ frontend/src/App.tsx
@@ -1,0 +1,11 @@
+import { Text } from '@mantine/core'
+
+function App() {
+	return (
+		<div className='App'>
+			<Text>Hello world!</Text>
+		</div>
+	)
+}
+
+export default App
