+++ frontend/src/index.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import { MantineProvider } from '@mantine/core'
+import App from './App'
+import reportWebVitals from './reportWebVitals'
+
+const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
+root.render(
+	<React.StrictMode>
+		<MantineProvider withGlobalStyles withNormalizeCSS>
+			<App />
+		</MantineProvider>
+	</React.StrictMode>
+)
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals()
